update: 添加问题“2975.移除栅栏得到的正方形田地的最大面积”的代码和题解 (#1331) 2942: AC.cpp(#1326) - AC,100.00%,84.62% + en
docs: en(2026.1.14+)2026.1.15
