+++ README.md
@@ -963,6 +963,7 @@
 |2965.找出缺失和重复的数字|简单|<a href="https://leetcode.cn/problems/find-missing-and-repeated-values/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2024/05/31/LeetCode%202965.%E6%89%BE%E5%87%BA%E7%BC%BA%E5%A4%B1%E5%92%8C%E9%87%8D%E5%A4%8D%E7%9A%84%E6%95%B0%E5%AD%97/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/139357662" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/find-missing-and-repeated-values/solutions/2796819/letmefly-2965zhao-chu-que-shi-he-zhong-f-p2rn/" target="_blank">LeetCode题解</a>|
 |2970.统计移除递增子数组的数目I|简单|<a href="https://leetcode.cn/problems/count-the-number-of-incremovable-subarrays-i/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2024/07/10/LeetCode%202970.%E7%BB%9F%E8%AE%A1%E7%A7%BB%E9%99%A4%E9%80%92%E5%A2%9E%E5%AD%90%E6%95%B0%E7%BB%84%E7%9A%84%E6%95%B0%E7%9B%AEI/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/140310063" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/count-the-number-of-incremovable-subarrays-i/solutions/2836851/letmefly-2970tong-ji-yi-chu-di-zeng-zi-s-3d2z/" target="_blank">LeetCode题解</a>|
 |2974.最小数字游戏|简单|<a href="https://leetcode.cn/problems/minimum-number-game/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2024/07/12/LeetCode%202974.%E6%9C%80%E5%B0%8F%E6%95%B0%E5%AD%97%E6%B8%B8%E6%88%8F/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/140365205" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/minimum-number-game/solutions/2840437/letmefly-2974zui-xiao-shu-zi-you-xi-pai-ib5em/" target="_blank">LeetCode题解</a>|
+|2975.移除栅栏得到的正方形田地的最大面积|中等|<a href="https://leetcode.cn/problems/maximum-square-area-by-removing-fences-from-a-field/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2026/01/17/LeetCode%202975.%E7%A7%BB%E9%99%A4%E6%A0%85%E6%A0%8F%E5%BE%97%E5%88%B0%E7%9A%84%E6%AD%A3%E6%96%B9%E5%BD%A2%E7%94%B0%E5%9C%B0%E7%9A%84%E6%9C%80%E5%A4%A7%E9%9D%A2%E7%A7%AF/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/157068738" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/maximum-square-area-by-removing-fences-from-a-field/solutions/3882988/letmefly-2975yi-chu-zha-lan-de-dao-de-zh-wua2/" target="_blank">LeetCode题解</a>|
 |2982.找出出现至少三次的最长特殊子字符串II|中等|<a href="https://leetcode.cn/problems/find-longest-special-substring-that-occurs-thrice-ii/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2024/05/30/LeetCode%202982.%E6%89%BE%E5%87%BA%E5%87%BA%E7%8E%B0%E8%87%B3%E5%B0%91%E4%B8%89%E6%AC%A1%E7%9A%84%E6%9C%80%E9%95%BF%E7%89%B9%E6%AE%8A%E5%AD%90%E5%AD%97%E7%AC%A6%E4%B8%B2II/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/139334864" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/find-longest-special-substring-that-occurs-thrice-ii/solutions/2795996/letmefly-2982zhao-chu-chu-xian-zhi-shao-9ubg1/" target="_blank">LeetCode题解</a>|
 |2999.统计强大整数的数目|困难|<a href="https://leetcode.cn/problems/count-the-number-of-powerful-integers/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/04/12/LeetCode%202999.%E7%BB%9F%E8%AE%A1%E5%BC%BA%E5%A4%A7%E6%95%B4%E6%95%B0%E7%9A%84%E6%95%B0%E7%9B%AE/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/147191672" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/count-the-number-of-powerful-integers/solutions/3649695/letmefly-2999tong-ji-qiang-da-zheng-shu-jnya8/" target="_blank">LeetCode题解</a>|
 |3000.对角线最长的矩形的面积|简单|<a href="https://leetcode.cn/problems/maximum-area-of-longest-diagonal-rectangle/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/08/26/LeetCode%203000.%E5%AF%B9%E8%A7%92%E7%BA%BF%E6%9C%80%E9%95%BF%E7%9A%84%E7%9F%A9%E5%BD%A2%E7%9A%84%E9%9D%A2%E7%A7%AF/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/150871371" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/maximum-area-of-longest-diagonal-rectangle/solutions/3764586/letmefly-3000dui-jiao-xian-zui-chang-de-cxvkc/" target="_blank">LeetCode题解</a>|
